增加权限

+++ trunk/src/main/java/com/zbsjk/controller/UserController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.zbsjk.ext.SecurityException;
 import com.zbsjk.model.entity.UserInfo;
 import com.zbsjk.model.vo.PutpwdVo;
 import com.zbsjk.service.UserService;
@@ -26,6 +27,10 @@ public class UserController {
 	@RequestMapping(value ="/user", method = RequestMethod.POST, produces = "application/json; charset=utf-8")
 	public Object addUser(HttpServletRequest request, HttpServletResponse response,
 			@RequestBody UserInfo userInfo){
+		UserInfo user = (UserInfo)request.getSession().getAttribute("user");
+		if(null==user){
+			throw new SecurityException("user", "请先登录");
+		}
 		return userService.addUser(userInfo);
 	}
 	
@@ -33,13 +38,22 @@ public class UserController {
 	public Object updateUser(HttpServletRequest request, HttpServletResponse response,
 			@PathVariable Integer userId,
 			@RequestBody UserInfo userInfo){
+		UserInfo user = (UserInfo)request.getSession().getAttribute("user");
+		if(null==user){
+			throw new SecurityException("user", "请先登录");
+		}
 		userInfo.setUserId(userId);
+		userInfo.setUserPwd(null);
 		return userService.updateUser(userInfo);
 	}
 	
 	@RequestMapping(value ="/user/{userId}", method = RequestMethod.DELETE, produces = "application/json; charset=utf-8")
 	public Object deleteUser(HttpServletRequest request, HttpServletResponse response,
 			@PathVariable Integer userId){
+		UserInfo user = (UserInfo)request.getSession().getAttribute("user");
+		if(null==user){
+			throw new SecurityException("user", "请先登录");
+		}
 		return userService.deleteUser(userId);
 	}
 	
@@ -48,12 +62,31 @@ public class UserController {
 			UserInfo userInfo,
 			@RequestParam Integer pageNo,
 			@RequestParam Integer pageSize){
+		UserInfo user = (UserInfo)request.getSession().getAttribute("user");
+		if(null==user){
+			throw new SecurityException("user", "请先登录");
+		}
+		
+		if(user.getRoleId().equals(2)){
+			userInfo.setUserCity(user.getUserCity());
+		}else if(user.getRoleId().equals(4)){
+			userInfo.setUserCity(user.getUserCity());
+			userInfo.setUserArea(user.getUserArea());
+		}else if(user.getRoleId().equals(1)){
+			
+		}else{
+			throw new SecurityException("user", "无权限");
+		}
 		return userService.getUserList(userInfo,pageNo,pageSize);
 	}
 	
 	@RequestMapping(value ="/user/{userid}/resetpwd", method = RequestMethod.PUT, produces = "application/json; charset=utf-8")
 	public Object resetpwd(HttpServletRequest request, HttpServletResponse response,
 			@PathVariable Integer userId){
+		UserInfo user = (UserInfo)request.getSession().getAttribute("user");
+		if(null==user){
+			throw new SecurityException("user", "请先登录");
+		}
 		return userService.resetpwd(userId);
 	}
 	
@@ -61,6 +94,10 @@ public class UserController {
 	public Object putpwd(HttpServletRequest request, HttpServletResponse response,
 			@PathVariable Integer userId,
 			@RequestBody PutpwdVo putpwdVo){
+		UserInfo user = (UserInfo)request.getSession().getAttribute("user");
+		if(null==user){
+			throw new SecurityException("user", "请先登录");
+		}
 		return userService.putpwd(putpwdVo,userId);
 	}
